avoid importing tornado in the base async socket implementation
